Revert "Flush messages before DeathWatchNotification, #28695 (#28940)"
This reverts commit f6ceb4d49aeb7ab28e9e7bd2963695b3a539c3af.
