--- akka-actor/src/main/mima-filters/2.6.6.backwards.excludes/28695-isTerminating.excludes
@@ -1,2 +1,0 @@
-# #28695 added isTerminated to ExtendedActorSystem
-ProblemFilters.exclude[ReversedMissingMethodProblem]("akka.actor.ExtendedActorSystem.isTerminating")
