--- akka-remote/src/main/mima-filters/2.6.6.backwards.excludes/28695-watch-flush.excludes
@@ -1,6 +1,0 @@
-# #28695 flush before sending DeathWatchNotification
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.remote.artery.FlushOnShutdown.props")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.remote.artery.FlushOnShutdown.timeoutTask")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.remote.artery.FlushOnShutdown.this")
-ProblemFilters.exclude[FinalClassProblem]("akka.remote.artery.ActorSystemTerminating")
-ProblemFilters.exclude[FinalClassProblem]("akka.remote.artery.ActorSystemTerminatingAck")
