+++ akka-remote/src/main/resources/reference.conf
@@ -980,12 +980,6 @@ akka {
         # remote messages has been completed
         shutdown-flush-timeout = 1 second
 
-        # Before sending notificaiton of terminated actor (DeathWatchNotification) other messages
-        # will be flushed to make sure that the Terminated message arrives after other messages.
-        # It will wait this long for the flush acknowledgement before continuing.
-        # The flushing can be disabled by setting this to `off`.
-        death-watch-notification-flush-timeout = 3 seconds
-
         # See 'inbound-max-restarts'
         inbound-restart-timeout = 5 seconds
