+++ akka-remote/src/main/scala/akka/remote/artery/ArterySettings.scala
@@ -152,18 +152,7 @@ private[akka] final class ArterySettings private (config: Config) {
     val ShutdownFlushTimeout: FiniteDuration =
       config
         .getMillisDuration("shutdown-flush-timeout")
-        .requiring(timeout => timeout > Duration.Zero, "shutdown-flush-timeout must be more than zero")
-    val DeathWatchNotificationFlushTimeout: FiniteDuration = {
-      toRootLowerCase(config.getString("death-watch-notification-flush-timeout")) match {
-        case "off" => Duration.Zero
-        case _ =>
-          config
-            .getMillisDuration("death-watch-notification-flush-timeout")
-            .requiring(
-              interval => interval > Duration.Zero,
-              "death-watch-notification-flush-timeout must be more than zero, or off")
-      }
-    }
+        .requiring(interval => interval > Duration.Zero, "shutdown-flush-timeout must be more than zero")
     val InboundRestartTimeout: FiniteDuration =
       config
         .getMillisDuration("inbound-restart-timeout")
