+++ akka-remote/src/main/scala/akka/remote/artery/Control.scala
@@ -11,7 +11,6 @@ import scala.concurrent.Promise
 import scala.util.Try
 
 import akka.Done
-import akka.annotation.InternalApi
 import akka.event.Logging
 import akka.remote.UniqueAddress
 import akka.stream.Attributes
@@ -22,13 +21,11 @@ import akka.stream.stage._
 import akka.util.OptionVal
 
 /** INTERNAL API: marker trait for protobuf-serializable artery messages */
-@InternalApi
 private[remote] trait ArteryMessage extends Serializable
 
 /**
  * INTERNAL API: Marker trait for reply messages
  */
-@InternalApi
 private[remote] trait Reply extends ControlMessage
 
 /**
@@ -36,43 +33,26 @@ private[remote] trait Reply extends ControlMessage
  * Marker trait for control messages that can be sent via the system message sub-channel
  * but don't need full reliable delivery. E.g. `HandshakeReq` and `Reply`.
  */
-@InternalApi
 private[remote] trait ControlMessage extends ArteryMessage
 
 /**
  * INTERNAL API
  */
-@InternalApi
 private[remote] final case class Quarantined(from: UniqueAddress, to: UniqueAddress) extends ControlMessage
 
 /**
  * INTERNAL API
  */
-@InternalApi
-private[remote] final case class ActorSystemTerminating(from: UniqueAddress) extends ControlMessage
+private[remote] case class ActorSystemTerminating(from: UniqueAddress) extends ControlMessage
 
 /**
  * INTERNAL API
  */
-@InternalApi
-private[remote] final case class ActorSystemTerminatingAck(from: UniqueAddress) extends ArteryMessage
+private[remote] case class ActorSystemTerminatingAck(from: UniqueAddress) extends ArteryMessage
 
 /**
  * INTERNAL API
  */
-@InternalApi
-private[remote] case object Flush extends ControlMessage
-
-/**
- * INTERNAL API
- */
-@InternalApi
-private[remote] case object FlushAck extends ArteryMessage
-
-/**
- * INTERNAL API
- */
-@InternalApi
 private[remote] object InboundControlJunction {
 
   /**
@@ -107,7 +87,6 @@ private[remote] object InboundControlJunction {
 /**
  * INTERNAL API
  */
-@InternalApi
 private[remote] class InboundControlJunction
     extends GraphStageWithMaterializedValue[
       FlowShape[InboundEnvelope, InboundEnvelope],
@@ -171,7 +150,6 @@ private[remote] class InboundControlJunction
 /**
  * INTERNAL API
  */
-@InternalApi
 private[remote] object OutboundControlJunction {
   private[remote] trait OutboundControlIngress {
     def sendControlMessage(message: ControlMessage): Unit
@@ -181,7 +159,6 @@ private[remote] object OutboundControlJunction {
 /**
  * INTERNAL API
  */
-@InternalApi
 private[remote] class OutboundControlJunction(
     outboundContext: OutboundContext,
     outboundEnvelopePool: ObjectPool[ReusableOutboundEnvelope])
